+++ _archive/FilteredSampleProd.py
@@ -1,0 +1,31 @@
+{
+  "id": 40,
+  "sku": "Cup-2",
+  "ebay_account_id": 1,
+  "title": "Copeland Spode Queen Elizabeth II 1953 Coronation Mug Tankard",
+  "description": "This is a Copeland Spode mug commemorating the coronation of Queen Elizabeth II on June 2nd, 1953. The mug has a creamy white glaze with brown transfer printing and coloured crests. It features a portrait of Queen Elizabeth II, the royal crest, and text marking the occasion.",
+  "condition_description": "Used condition with staining and minor chips around the rims. The glaze is crazed and discolored. There are marks and scuffs on the exterior surface.",
+  "suggested_price": 12,
+  "selected_category_path": "Historical Memorabilia >> Collectables >> Royalty Collectables",
+  "selected_ebay_category_id": "39630",
+  "images": [
+    {
+      "b2_url": "https://eBayImages.s3.us-east-005.backblazeb2.com/ebay_images/Cup-2/Cup-2-1_81.jpg"
+    },
+    {
+      "b2_url": "https://eBayImages.s3.us-east-005.backblazeb2.com/ebay_images/Cup-2/Cup-2-1_78.jpg"
+    },
+    {
+      "b2_url": "https://eBayImages.s3.us-east-005.backblazeb2.com/ebay_images/Cup-2/Cup-2-1_80.jpg"
+    },
+    {
+      "b2_url": "https://eBayImages.s3.us-east-005.backblazeb2.com/ebay_images/Cup-2/Cup-2-1_82.jpg"
+    },
+    {
+      "b2_url": "https://eBayImages.s3.us-east-005.backblazeb2.com/ebay_images/Cup-2/Cup-2-1_79.jpg"
+    },
+    {
+      "b2_url": "https://eBayImages.s3.us-east-005.backblazeb2.com/ebay_images/Cup-2/Cup-2-1_77.jpg"
+    }
+  ]
+}