+++ start_mcp_inspector.sh
@@ -1,8 +1,8 @@
 
-print "Killing any existing FastMCP inspector processes..."
+echo "Killing any existing FastMCP inspector processes..."
 lsof -ti :6277 | xargs kill -9
 lsof -ti :6274 | xargs   kill -9
 
-echo "Starting FastMCP inspector..."
-npx @modelcontextprotocol/inspector python /Users/petetreadaway/Projects/ebay-mcp-server/src/main_server.py
+echo "Starting eBayAPI-MCP-Server inspector..."
+npx @modelcontextprotocol/inspector --config /Users/petetreadaway/Projects/ebay-mcp-server/MCP_Inspector_Config.json --server ebayAPI-mcp-server